pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.0",
    "isort>=8.0.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
]
